Backend/requirements.txt
Flask
flask-cors
python-dotenv
pymongo
pandas
scikit-learn
xgboost
shap
requests
gunicorn
google-generativeai